freecodecamp-estrutura-dados/src/programacao/algoritmos/intermediarios/parte1/05-spinal-case/05-spinalCase.spec.ts
import { spinalCase } from './05-spinalCase';

describe('Deve retornar uma string Hifenizada quando palavras', () => {
  test('Separadas por espaços', () => {
    expect(spinalCase("This Is Spinal Tap"))
      .toEqual('this-is-spinal-tap');
  })
  test('Juntas - grudadas', () => {
    expect(spinalCase("ThisIsSpinalTap"))
      .toEqual('this-is-spinal-tap');
  })
  test('Com Underscore', () => {
    expect(spinalCase("This_Is_Spinal_Tap"))
      .toEqual('this-is-spinal-tap');
  })
  test('Com Underscore', () => {
    expect(spinalCase("Teletubbies say Eh-oh"))
      .toEqual('teletubbies-say-eh-oh');
  })
})